--transfer

+++ auth/src/main/java/de/mein/auth/data/access/CertificateManager.java
@@ -5,7 +5,6 @@ import de.mein.auth.data.db.dao.CertificateDao;
 import de.mein.auth.tools.Cryptor;
 import de.mein.sql.ISQLQueries;
 import de.mein.sql.SqlQueriesException;
-
 import org.bouncycastle.jce.X509Principal;
 import org.bouncycastle.jce.provider.BouncyCastleProvider;
 import org.bouncycastle.x509.X509V3CertificateGenerator;
@@ -14,7 +13,6 @@ import javax.crypto.BadPaddingException;
 import javax.crypto.IllegalBlockSizeException;
 import javax.crypto.NoSuchPaddingException;
 import javax.net.ssl.*;
-
 import java.io.*;
 import java.math.BigInteger;
 import java.net.ServerSocket;
@@ -66,9 +64,11 @@ public class CertificateManager extends FileRelatedManager {
         //load keystore
         keyStoreFile = new File(createWorkingPath() + KS_FILENAME);
         // we want to start with a clean keystore
-        boolean deleted = keyStoreFile.delete();
-        if (!deleted)
-            System.err.println("CertificateManager().KEYSTORE.NOT.DELETED");
+        if (keyStoreFile.exists()) {
+            boolean deleted = keyStoreFile.delete();
+            if (!deleted)
+                System.err.println("CertificateManager().KEYSTORE.NOT.DELETED");
+        }
         keyStore = KeyStore.getInstance(KeyStore.getDefaultType());
         keyStore.load(null, PASS.toCharArray());
         //init DB stuff
